1.4.4

+++ internal/server/models.go
@@ -9,5 +9,5 @@ type Gauge struct {
 type Counter struct {
 	Metric string
 	Name   string
-	Value  float64
+	Value  int
 }
